medical-coding-app/requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
python-dotenv==1.0.0
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
quickumls==1.4.2
spacy==3.4.4
requests==2.31.0
Werkzeug==2.3.7
nltk==3.8.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

pytest
pytest-mock